refactor: Relocate `PixelContext` to the core pipelines module
Moves the `PixelContext` definition from the interfaces layer to the internal pipelines logic. This consolidates generation-specific context structures and maintains a cleaner separation between public-facing API definitions and internal pipeline state management.
