+++ max/python/max/interfaces/__init__.py
@@ -50,7 +50,6 @@ from .pipeline_variants import (
     EmbeddingsGenerationOutput,
     ImageContentPart,
     ImageMetadata,
-    PixelContext,
     PixelGenerationInputs,
     PixelGenerationOutput,
     PixelGenerationRequest,
@@ -132,7 +131,6 @@ __all__ = [
     "PipelineTask",
     "PipelineTokenizer",
     "PipelinesFactory",
-    "PixelContext",
     "PixelGenerationInputs",
     "PixelGenerationOutput",
     "PixelGenerationRequest",
