+++ max/python/max/interfaces/pipeline_variants/pixel_generation.py
@@ -19,16 +19,15 @@ responses, including status tracking and pixel data encapsulation.
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Any, Generic
+from typing import Generic
 
 import msgspec
 import numpy as np
 import numpy.typing as npt
-from max.interfaces.context import BaseContext
 from max.interfaces.pipeline import PipelineInputs, PipelineOutput
 from max.interfaces.request import Request, RequestID
 from max.interfaces.status import GenerationStatus
-from max.interfaces.tokens import TokenBuffer
+from max.pipelines import PixelContext
 from typing_extensions import TypeVar
 
 
@@ -88,155 +87,6 @@ class PixelGenerationRequest(Request):
     def __post_init__(self) -> None:
         if self.prompt == "":
             raise ValueError("Prompt must be provided.")
-
-
-@dataclass(kw_only=True)
-class PixelContext(BaseContext):
-    """A model-ready context for image/video generation requests.
-
-    Per the design doc, this class contains only numeric data that the model
-    will execute against. User-facing strings (prompt, negative_prompt) are
-    consumed during tokenization and do not appear here.
-
-    All preprocessing is performed by PixelGenerationTokenizer.new_context():
-    - Prompt tokenization -> tokens field
-    - Negative prompt tokenization -> negative_tokens field
-    - Timestep schedule computation -> timesteps field
-    - Initial noise generation -> initial_noise field
-
-    Configuration:
-        tokens: Tokenized prompt IDs (TokenBuffer).
-        request_id: A unique identifier for this generation request.
-        negative_tokens: Tokenized negative prompt IDs (TokenBuffer).
-        timesteps: Precomputed timestep schedule for denoising.
-        initial_noise: Precomputed initial noise (latents).
-        height: Height of the generated image/video in pixels.
-        width: Width of the generated image/video in pixels.
-        num_inference_steps: Number of denoising steps.
-        guidance_scale: Guidance scale for classifier-free guidance.
-        num_images_per_prompt: Number of images/videos to generate per prompt.
-        model_name: Name of the model being used.
-    """
-
-    # Tokenized prompts
-    tokens: TokenBuffer
-    """Primary encoder tokens."""
-
-    # Request identification
-    request_id: RequestID = field(default_factory=RequestID)
-
-    model_name: str = field(default="")
-
-    mask: npt.NDArray[np.bool_] | None = field(default=None)
-    """Mask for text encoder's attention."""
-
-    tokens_2: TokenBuffer | None = field(default=None)
-    """Secondary encoder tokens. None for single-encoder models."""
-
-    negative_tokens: TokenBuffer | None = field(default=None)
-    """Negative tokens for primary encoder."""
-
-    negative_tokens_2: TokenBuffer | None = field(default=None)
-    """Negative tokens for secondary encoder. None for single-encoder models."""
-
-    extra_params: dict[str, npt.NDArray[Any]] = field(default_factory=dict)
-    """Model-specific numeric parameters (e.g., cfg_normalization values)."""
-
-    # Precomputed tensors
-    timesteps: npt.NDArray[np.float32] = field(
-        default_factory=lambda: np.array([], dtype=np.float32)
-    )
-    """Precomputed timesteps schedule for denoising."""
-
-    sigmas: npt.NDArray[np.float32] = field(
-        default_factory=lambda: np.array([], dtype=np.float32)
-    )
-    """Precomputed sigmas schedule for denoising."""
-
-    latents: npt.NDArray[np.float32] = field(
-        default_factory=lambda: np.array([], dtype=np.float32)
-    )
-    """Precomputed initial noise (latents) for generation."""
-
-    latent_image_ids: npt.NDArray[np.float32] = field(
-        default_factory=lambda: np.array([], dtype=np.float32)
-    )
-    """Precomputed latent image IDs for generation."""
-
-    # Image generation parameters
-    height: int = field(default=1024)
-    width: int = field(default=1024)
-    num_inference_steps: int = field(default=50)
-    guidance_scale: float = field(default=3.5)
-    guidance: npt.NDArray[np.float32] | None = None
-    true_cfg_scale: float = field(default=1.0)
-    num_warmup_steps: int = field(default=0)
-    num_images_per_prompt: int = field(default=1)
-
-    # Generation status
-    _status: GenerationStatus = field(default=GenerationStatus.ACTIVE)
-
-    @property
-    def status(self) -> GenerationStatus:
-        """Current generation status of the request."""
-        return self._status
-
-    @status.setter
-    def status(self, value: GenerationStatus) -> None:
-        """Update the generation status."""
-        self._status = value
-
-    @property
-    def is_done(self) -> bool:
-        """Whether the request has completed generation."""
-        return self._status.is_done
-
-    @property
-    def needs_ce(self) -> bool:
-        """Whether this context needs context encoding.
-
-        For image generation, we never need context encoding since
-        we process the full prompt at once through the text encoder.
-        """
-        return False
-
-    @property
-    def active_length(self) -> int:
-        """Current sequence length for batch constructor compatibility."""
-        return 1
-
-    @property
-    def current_length(self) -> int:
-        """Current length for batch constructor compatibility."""
-        return 1
-
-    @property
-    def processed_length(self) -> int:
-        """Processed length for batch constructor compatibility."""
-        return 0
-
-    def compute_num_available_steps(self, max_seq_len: int) -> int:
-        """Compute number of available steps for scheduler compatibility.
-
-        For image generation, this returns the number of inference steps.
-        """
-        return self.num_inference_steps
-
-    def reset(self) -> None:
-        """Resets the context's state."""
-        self.status = GenerationStatus.ACTIVE
-
-    def update(self, latents: npt.NDArray[Any]) -> None:
-        """Update the context with newly generated latents/image data."""
-        self.latents = latents
-
-    def to_generation_output(self) -> PixelGenerationOutput:
-        """Convert this context to a PixelGenerationOutput object."""
-        return PixelGenerationOutput(
-            request_id=self.request_id,
-            final_status=self.status,
-            pixel_data=self.latents,
-        )
 
 
 PixelGenerationContextType = TypeVar(
